requirements.txt
arxiv==2.1.3
python-dotenv==1.0.1
boto3==1.35.36